add commands, fix comment sampling bug
The "comment sampling bug" was that non-instrument data (i.e. comments) was sampled, too.
Adding a comment in a long or complicated song may be desirable to add meta-information, just like bar lines in sheet music sometimes include a a running index. Of course, marking the current bar line this should not generate a pause; therefore, comments as well as commands should rather be skipped.