fix bug where omitting `--bpm` would ignore first tempo declaration
Further, this commit covers the case where a change in tempo occurs later in a roll, without the composer *starting* the roll with a tempo declaration.

This commit also cleans up the code by removing an unnecessary closure (thus saving an indentation level), choosing `else if` instead of `if` where appropriate and giving variables more consistent naming.